changes in search bar components
