--- src/common/components/search-element/components/index.ts
@@ -1,1 +1,0 @@
-export { SearchInput } from './search-input/index'
