--- src/common/components/search-element/helpers/get-dropdown-options-format.ts
@@ -1,12 +1,0 @@
-import { type DropdownOption, type Company } from "~/common/types";
-
-const getDropdownOptionsFormat = (items: Company[]): DropdownOption[] => {
-  const mappedItems = items.map((company) => ({
-    value: company.id,
-    label: company.name
-  }));
-
-  return mappedItems
-}
-
-export { getDropdownOptionsFormat };
