--- src/common/components/search-element/helpers/index.ts
@@ -1,2 +1,0 @@
-export { getDropdownOptionsFormat } from './get-dropdown-options-format';
-export { mapCompanies } from './map-companies';
