--- src/common/types/company/get-companies-response-dto.type.ts
@@ -1,10 +1,0 @@
-import { type Company } from "./company.type";
-
-type GetCompaniesResponse = {
-  count: number;
-  next: string | null;
-  previous: string | null;
-  results: Company[];
-};
-
-export { type GetCompaniesResponse };
