+++ src/pages/home-page/components/main-content/components/index.ts
@@ -1,1 +1,2 @@
 export { BanerBlock } from './baner-block/index';
+export { SearchBlock } from './search-block/index';
