+++ src/pages/home-page/components/main-content/components/search-block/components/index.ts
@@ -1,0 +1,2 @@
+export { NotFound } from './not-found/index';
+export { SearchElement } from './search-element/index'
