+++ src/pages/home-page/components/main-content/components/search-block/components/search-element/components/search-input/styles.module.scss
