+++ src/pages/home-page/components/main-content/components/search-block/components/search-element/styles.module.scss
