+++ src/pages/home-page/components/main-content/components/search-block/styles.module.scss
@@ -1,0 +1,3 @@
+.search_block {
+  width: 100%;
+}
