+++ src/pages/home-page/components/main-content/index.tsx
@@ -1,5 +1,4 @@
-import { SearchElement } from '~/common/components';
-import { BanerBlock } from './components/index';
+import { BanerBlock, SearchBlock } from './components/index';
 import styles from './styles.module.scss';
 
 const MainContent = () => {
@@ -8,7 +7,7 @@ const MainContent = () => {
     <div className={styles['main_content_wrapper']}>
       <div className={styles['main_content']}>
         <BanerBlock />
-        <SearchElement />
+        <SearchBlock />
       </div>
     </div>
   )
